requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
Deprecated==1.2.14
Django==4.2.3
django-environ==0.10.0
examples==1.0.2
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
kombu==5.3.1
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pydantic==2.1.1
Pygments==2.15.1
pymodbus==3.4.1
pyparsing==3.1.1
pyserial==3.5
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
zope.interface==6.0
djangorestframework~=3.14.0
django-filter~=23.2